Set manifestation and expression dates for components.
Closes #55
