+++ tests/test_structured_document.py
@@ -2,9 +2,7 @@ from unittest import TestCase
 from nose.tools import *  # noqa
 from datetime import date
 
-from lxml.etree import LxmlSyntaxError
-
-from cobalt import Act, AmendmentEvent, RepealEvent, Judgment, datestring
+from cobalt import Act, datestring
 from cobalt.schemas import assert_validates
 
 
@@ -184,235 +182,6 @@ class StructuredDocumentTestCase(TestCase):
         assert_in("Expected act as first child of root element, but got somethingElse instead",
                   raised.exception.args)
 
-    def test_components(self):
-        a = Act(xml="""<?xml version="1.0" encoding="UTF-8"?>
-<akomaNtoso xmlns="http://docs.oasis-open.org/legaldocml/ns/akn/3.0" xsi:schemaLocation="http://docs.oasis-open.org/legaldocml/akn-core/v1.0/os/part2-specs/schemas/akomantoso30.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <act contains="singleVersion" name="act">
-    <meta>
-      <identification source="#slaw">
-        <FRBRWork>
-          <FRBRthis value="/na/act/1977/25/!main"/>
-          <FRBRuri value="/na/act/1977/25"/>
-          <FRBRalias value="Livestock Improvement Act, 1977" name="title"/>
-          <FRBRdate date="1977-03-23" name="Generation"/>
-          <FRBRauthor href="#council"/>
-          <FRBRcountry value="na"/>
-        </FRBRWork>
-        <FRBRExpression>
-          <FRBRthis value="/na/act/1977/25/eng@1993-12-02/!main"/>
-          <FRBRuri value="/na/act/1977/25/eng@1993-12-02"/>
-          <FRBRdate date="1993-12-02" name="Generation"/>
-          <FRBRauthor href="#council"/>
-          <FRBRlanguage language="eng"/>
-        </FRBRExpression>
-        <FRBRManifestation>
-          <FRBRthis value="/na/act/1977/25/eng@1993-12-02/!main"/>
-          <FRBRuri value="/na/act/1977/25/eng@1993-12-02"/>
-          <FRBRdate date="2020-03-25" name="Generation"/>
-          <FRBRauthor href="#slaw"/>
-        </FRBRManifestation>
-      </identification>
-      <publication number="5462" name="South African Government Gazette" showAs="South African Government Gazette" date="1977-03-23"/>
-    </meta>
-    <body>
-      <section eId="section_1">
-        <content>
-          <p></p>
-        </content>
-      </section>
-    </body>
-    <attachments>
-      <attachment eId="att_1">
-        <heading>Schedule</heading>
-        <doc name="schedule">
-          <meta>
-            <identification source="#slaw">
-              <FRBRWork>
-                <FRBRthis value="/na/act/1977/25/!schedule-A"/>
-                <FRBRuri value="/na/act/1977/25"/>
-                <FRBRalias value="Schedule" name="title"/>
-                <FRBRdate date="1980-01-01" name="Generation"/>
-                <FRBRauthor href="#council"/>
-                <FRBRcountry value="na"/>
-              </FRBRWork>
-              <FRBRExpression>
-                <FRBRthis value="/na/act/1977/25/eng@1993-12-02/!schedule-A"/>
-                <FRBRuri value="/na/act/1977/25/eng@1993-12-02"/>
-                <FRBRdate date="1980-01-01" name="Generation"/>
-                <FRBRauthor href="#council"/>
-                <FRBRlanguage language="eng"/>
-              </FRBRExpression>
-              <FRBRManifestation>
-                <FRBRthis value="/na/act/1977/25/eng@1993-12-02/!schedule-A"/>
-                <FRBRuri value="/na/act/1977/25/eng@1993-12-02"/>
-                <FRBRdate date="2020-03-25" name="Generation"/>
-                <FRBRauthor href="#slaw"/>
-              </FRBRManifestation>
-            </identification>
-          </meta>
-          <mainBody>
-            <paragraph eId="paragraph_1">
-              <content>
-                <p>This is the content of the Schedule!</p>
-              </content>
-            </paragraph>
-          </mainBody>
-        </doc>
-      </attachment>
-    </attachments>
-    <components>
-      <component eId="comp_1">
-        <heading>Schedule</heading>
-        <doc name="schedule">
-          <meta>
-            <identification source="#slaw">
-              <FRBRWork>
-                <FRBRthis value="/na/act/1977/25/!schedule-XXX"/>
-                <FRBRuri value="/na/act/1977/25"/>
-                <FRBRalias value="Schedule" name="title" />
-                <FRBRdate date="1980-01-01" name="Generation"/>
-                <FRBRauthor href="#council"/>
-                <FRBRcountry value="na"/>
-              </FRBRWork>
-              <FRBRExpression>
-                <FRBRthis value="/na/act/1977/25/eng@1993-12-02/!schedule-XXX"/>
-                <FRBRuri value="/na/act/1977/25/eng@1993-12-02"/>
-                <FRBRdate date="1980-01-01" name="Generation"/>
-                <FRBRauthor href="#council"/>
-                <FRBRlanguage language="eng"/>
-              </FRBRExpression>
-              <FRBRManifestation>
-                <FRBRthis value="/na/act/1977/25/eng@1993-12-02/!schedule-XXX"/>
-                <FRBRuri value="/na/act/1977/25/eng@1993-12-02"/>
-                <FRBRdate date="2020-03-25" name="Generation"/>
-                <FRBRauthor href="#slaw"/>
-              </FRBRManifestation>
-            </identification>
-           </meta>
-          <mainBody>
-            <paragraph eId="paragraph_1">
-              <content>
-                <p>This is the content of the Schedule!</p>
-              </content>
-            </paragraph>
-          </mainBody>
-        </doc>
-      </component>
-    </components>
-  </act>
-</akomaNtoso>
-        """)
-        components = a.components()
-        self.assertEqual(['main', 'schedule-A', 'schedule-XXX'], sorted(components.keys()))
-        self.assertEqual('This is the content of the Schedule!',
-                         components['schedule-XXX'].doc.mainBody.paragraph.content.p)
-        assert_validates(a)
-
-        # this won't change any component names, because the existing names are used by default
-        frbr_uri = a.frbr_uri
-        frbr_uri.work_component = 'blah'
-        a.frbr_uri = frbr_uri
-        self.assertEqual(['main', 'schedule-A', 'schedule-XXX'], sorted(a.components().keys()))
-
-    def test_set_missing_component_with_default(self):
-        a = Act(xml="""<?xml version="1.0" encoding="UTF-8"?>
-        <akomaNtoso xmlns="http://docs.oasis-open.org/legaldocml/ns/akn/3.0" xsi:schemaLocation="http://docs.oasis-open.org/legaldocml/akn-core/v1.0/os/part2-specs/schemas/akomantoso30.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-          <act contains="singleVersion" name="act">
-            <meta>
-              <identification source="#slaw">
-                <FRBRWork>
-                  <FRBRthis value="/na/act/1977/25"/>
-                  <FRBRuri value="/na/act/1977/25"/>
-                  <FRBRalias value="Livestock Improvement Act, 1977" name="title"/>
-                  <FRBRdate date="1977-03-23" name="Generation"/>
-                  <FRBRauthor href="#council"/>
-                  <FRBRcountry value="na"/>
-                </FRBRWork>
-                <FRBRExpression>
-                  <FRBRthis value="/na/act/1977/25/eng@1993-12-02"/>
-                  <FRBRuri value="/na/act/1977/25/eng@1993-12-02"/>
-                  <FRBRdate date="1993-12-02" name="Generation"/>
-                  <FRBRauthor href="#council"/>
-                  <FRBRlanguage language="eng"/>
-                </FRBRExpression>
-                <FRBRManifestation>
-                  <FRBRthis value="/na/act/1977/25/eng@1993-12-02"/>
-                  <FRBRuri value="/na/act/1977/25/eng@1993-12-02"/>
-                  <FRBRdate date="2020-03-25" name="Generation"/>
-                  <FRBRauthor href="#slaw"/>
-                </FRBRManifestation>
-              </identification>
-              <publication number="5462" name="South African Government Gazette" showAs="South African Government Gazette" date="1977-03-23"/>
-            </meta>
-            <body>
-              <section eId="section_1">
-                <content>
-                  <p></p>
-                </content>
-              </section>
-            </body>
-          </act>
-        </akomaNtoso>
-                """)
-        self.assertEqual([None], sorted(a.components().keys()))
-        assert_validates(a)
-
-        # change the main component name implicitly
-        # should get 'main' as the default
-        frbr_uri = a.frbr_uri
-        frbr_uri.work_component = None
-        a.frbr_uri = frbr_uri
-        self.assertEqual(['main'], sorted(a.components().keys()))
-
-    def test_set_missing_component_explicitly(self):
-        a = Act(xml="""<?xml version="1.0" encoding="UTF-8"?>
-        <akomaNtoso xmlns="http://docs.oasis-open.org/legaldocml/ns/akn/3.0" xsi:schemaLocation="http://docs.oasis-open.org/legaldocml/akn-core/v1.0/os/part2-specs/schemas/akomantoso30.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-          <act contains="singleVersion" name="act">
-            <meta>
-              <identification source="#slaw">
-                <FRBRWork>
-                  <FRBRthis value="/na/act/1977/25"/>
-                  <FRBRuri value="/na/act/1977/25"/>
-                  <FRBRalias value="Livestock Improvement Act, 1977" name="title"/>
-                  <FRBRdate date="1977-03-23" name="Generation"/>
-                  <FRBRauthor href="#council"/>
-                  <FRBRcountry value="na"/>
-                </FRBRWork>
-                <FRBRExpression>
-                  <FRBRthis value="/na/act/1977/25/eng@1993-12-02"/>
-                  <FRBRuri value="/na/act/1977/25/eng@1993-12-02"/>
-                  <FRBRdate date="1993-12-02" name="Generation"/>
-                  <FRBRauthor href="#council"/>
-                  <FRBRlanguage language="eng"/>
-                </FRBRExpression>
-                <FRBRManifestation>
-                  <FRBRthis value="/na/act/1977/25/eng@1993-12-02"/>
-                  <FRBRuri value="/na/act/1977/25/eng@1993-12-02"/>
-                  <FRBRdate date="2020-03-25" name="Generation"/>
-                  <FRBRauthor href="#slaw"/>
-                </FRBRManifestation>
-              </identification>
-              <publication number="5462" name="South African Government Gazette" showAs="South African Government Gazette" date="1977-03-23"/>
-            </meta>
-            <body>
-              <section eId="section_1">
-                <content>
-                  <p></p>
-                </content>
-              </section>
-            </body>
-          </act>
-        </akomaNtoso>
-                """)
-        self.assertEqual([None], sorted(a.components().keys()))
-        assert_validates(a)
-
-        # change the main component name
-        frbr_uri = a.frbr_uri
-        frbr_uri.work_component = 'blah'
-        a.frbr_uri = frbr_uri
-        self.assertEqual(['blah'], sorted(a.components().keys()))
-
     def test_add_number(self):
         """ When adding an FRBRnumber element to a document that doesn't already have one, it
         must come after subtype.
@@ -460,266 +229,3 @@ class StructuredDocumentTestCase(TestCase):
         a.frbr_uri = '/na/act/p/1997/25/'
 
         assert_validates(a)
-
-
-class ActTestCase(TestCase):
-    maxDiff = None
-
-    def test_empty_act(self):
-        a = Act()
-        assert_equal(a.title, "Untitled")
-        assert_is_not_none(a.meta)
-        assert_is_not_none(a.body)
-
-        assert_validates(a)
-
-    def test_empty_body(self):
-        a = Act()
-        assert_not_equal(a.body.text, '')
-
-    def test_publication_date(self):
-        a = Act()
-        assert_is_none(a.publication_date)
-
-        a.publication_date = '2012-01-02'
-        assert_equal(datestring(a.publication_date), '2012-01-02')
-        assert_is_instance(a.publication_date, date)
-
-    def test_publication_number(self):
-        a = Act()
-        assert_is_none(a.publication_number)
-
-        a.publication_number = '1234'
-        assert_equal(a.publication_number, '1234')
-
-    def test_publication_name(self):
-        a = Act()
-        assert_is_none(a.publication_name)
-
-        a.publication_name = 'Publication'
-        assert_equal(a.publication_name, 'Publication')
-
-    def test_set_amendments(self):
-        a = Act()
-        a.frbr_uri = "/akn/za/act/1900/1"
-        a.amendments = [AmendmentEvent(date='2012-02-01', amending_uri='/za/act/1980/10', amending_title="Foo")]
-
-        xml = a.to_xml(encoding='unicode', pretty_print=True)
-        xml = xml.replace(datestring(date.today()), 'TODAY')
-
-        self.assertMultiLineEqual(
-            """<akomaNtoso xmlns="http://docs.oasis-open.org/legaldocml/ns/akn/3.0">
-  <act name="act" contains="singleVersion">
-    <meta>
-      <identification source="#cobalt">
-        <FRBRWork>
-          <FRBRthis value="/akn/za/act/1900/1/!main"/>
-          <FRBRuri value="/akn/za/act/1900/1"/>
-          <FRBRalias value="Untitled" name="title"/>
-          <FRBRdate date="1900" name="Generation"/>
-          <FRBRauthor href=""/>
-          <FRBRcountry value="za"/>
-          <FRBRnumber value="1"/>
-        </FRBRWork>
-        <FRBRExpression>
-          <FRBRthis value="/akn/za/act/1900/1/eng@TODAY/!main"/>
-          <FRBRuri value="/akn/za/act/1900/1/eng@TODAY"/>
-          <FRBRdate date="TODAY" name="Generation"/>
-          <FRBRauthor href=""/>
-          <FRBRlanguage language="eng"/>
-        </FRBRExpression>
-        <FRBRManifestation>
-          <FRBRthis value="/akn/za/act/1900/1/eng@TODAY/!main"/>
-          <FRBRuri value="/akn/za/act/1900/1/eng@TODAY"/>
-          <FRBRdate date="TODAY" name="Generation"/>
-          <FRBRauthor href=""/>
-        </FRBRManifestation>
-      </identification>
-      <lifecycle source="#cobalt">
-        <eventRef eId="amendment-2012-02-01" date="2012-02-01" type="amendment" source="#amendment-0-source"/>
-      </lifecycle>
-      <references source="#cobalt">
-        <TLCOrganization eId="cobalt" href="https://github.com/laws-africa/cobalt" showAs="cobalt"/>
-        <passiveRef eId="amendment-0-source" href="/za/act/1980/10" showAs="Foo"/>
-      </references>
-    </meta>
-    <body>
-      <section eId="sec_nn_1">
-        <content>
-          <p eId="sec_nn_1__p_1"/>
-        </content>
-      </section>
-    </body>
-  </act>
-</akomaNtoso>
-""", xml)
-
-        a.amendments = [
-            AmendmentEvent(date='2012-02-01', amending_uri='/za/act/1980/22', amending_title="Corrected"),
-            AmendmentEvent(date='2013-03-03', amending_uri='/za/act/1990/5', amending_title="Bar"),
-        ]
-
-        xml = a.to_xml(encoding='unicode', pretty_print=True)
-        xml = xml.replace(datestring(date.today()), 'TODAY')
-
-        self.assertMultiLineEqual(
-            """<akomaNtoso xmlns="http://docs.oasis-open.org/legaldocml/ns/akn/3.0">
-  <act name="act" contains="singleVersion">
-    <meta>
-      <identification source="#cobalt">
-        <FRBRWork>
-          <FRBRthis value="/akn/za/act/1900/1/!main"/>
-          <FRBRuri value="/akn/za/act/1900/1"/>
-          <FRBRalias value="Untitled" name="title"/>
-          <FRBRdate date="1900" name="Generation"/>
-          <FRBRauthor href=""/>
-          <FRBRcountry value="za"/>
-          <FRBRnumber value="1"/>
-        </FRBRWork>
-        <FRBRExpression>
-          <FRBRthis value="/akn/za/act/1900/1/eng@TODAY/!main"/>
-          <FRBRuri value="/akn/za/act/1900/1/eng@TODAY"/>
-          <FRBRdate date="TODAY" name="Generation"/>
-          <FRBRauthor href=""/>
-          <FRBRlanguage language="eng"/>
-        </FRBRExpression>
-        <FRBRManifestation>
-          <FRBRthis value="/akn/za/act/1900/1/eng@TODAY/!main"/>
-          <FRBRuri value="/akn/za/act/1900/1/eng@TODAY"/>
-          <FRBRdate date="TODAY" name="Generation"/>
-          <FRBRauthor href=""/>
-        </FRBRManifestation>
-      </identification>
-      <lifecycle source="#cobalt">
-        <eventRef eId="amendment-2012-02-01" date="2012-02-01" type="amendment" source="#amendment-0-source"/>
-        <eventRef eId="amendment-2013-03-03" date="2013-03-03" type="amendment" source="#amendment-1-source"/>
-      </lifecycle>
-      <references source="#cobalt">
-        <TLCOrganization eId="cobalt" href="https://github.com/laws-africa/cobalt" showAs="cobalt"/>
-        <passiveRef eId="amendment-0-source" href="/za/act/1980/22" showAs="Corrected"/>
-        <passiveRef eId="amendment-1-source" href="/za/act/1990/5" showAs="Bar"/>
-      </references>
-    </meta>
-    <body>
-      <section eId="sec_nn_1">
-        <content>
-          <p eId="sec_nn_1__p_1"/>
-        </content>
-      </section>
-    </body>
-  </act>
-</akomaNtoso>
-""", xml)
-
-        amendment = a.amendments[0]
-        assert_equal(datestring(amendment.date), '2012-02-01')
-        assert_equal(amendment.amending_uri, '/za/act/1980/22')
-        assert_equal(amendment.amending_title, 'Corrected')
-
-        amendment = a.amendments[1]
-        assert_equal(datestring(amendment.date), '2013-03-03')
-        assert_equal(amendment.amending_uri, '/za/act/1990/5')
-        assert_equal(amendment.amending_title, 'Bar')
-
-        assert_validates(a)
-
-    def test_set_repeal(self):
-        a = Act()
-        a.frbr_uri = "/akn/za/act/1900/1"
-        a.repeal = RepealEvent(date='2012-02-01', repealing_uri='/za/act/1980/10', repealing_title='Foo')
-        xml = a.to_xml(encoding='unicode', pretty_print=True)
-        xml = xml.replace(datestring(date.today()), 'TODAY')
-
-        self.assertMultiLineEqual(
-            """<akomaNtoso xmlns="http://docs.oasis-open.org/legaldocml/ns/akn/3.0">
-  <act name="act" contains="originalVersion">
-    <meta>
-      <identification source="#cobalt">
-        <FRBRWork>
-          <FRBRthis value="/akn/za/act/1900/1/!main"/>
-          <FRBRuri value="/akn/za/act/1900/1"/>
-          <FRBRalias value="Untitled" name="title"/>
-          <FRBRdate date="1900" name="Generation"/>
-          <FRBRauthor href=""/>
-          <FRBRcountry value="za"/>
-          <FRBRnumber value="1"/>
-        </FRBRWork>
-        <FRBRExpression>
-          <FRBRthis value="/akn/za/act/1900/1/eng@TODAY/!main"/>
-          <FRBRuri value="/akn/za/act/1900/1/eng@TODAY"/>
-          <FRBRdate date="TODAY" name="Generation"/>
-          <FRBRauthor href=""/>
-          <FRBRlanguage language="eng"/>
-        </FRBRExpression>
-        <FRBRManifestation>
-          <FRBRthis value="/akn/za/act/1900/1/eng@TODAY/!main"/>
-          <FRBRuri value="/akn/za/act/1900/1/eng@TODAY"/>
-          <FRBRdate date="TODAY" name="Generation"/>
-          <FRBRauthor href=""/>
-        </FRBRManifestation>
-      </identification>
-      <lifecycle source="#cobalt">
-        <eventRef eId="repeal-2012-02-01" date="2012-02-01" type="repeal" source="#repeal-source"/>
-      </lifecycle>
-      <references source="#cobalt">
-        <TLCOrganization eId="cobalt" href="https://github.com/laws-africa/cobalt" showAs="cobalt"/>
-        <passiveRef eId="repeal-source" href="/za/act/1980/10" showAs="Foo"/>
-      </references>
-    </meta>
-    <body>
-      <section eId="sec_nn_1">
-        <content>
-          <p eId="sec_nn_1__p_1"/>
-        </content>
-      </section>
-    </body>
-  </act>
-</akomaNtoso>
-""", xml)
-
-        assert_equal(a.repeal.repealing_uri, '/za/act/1980/10')
-        assert_equal(a.repeal.repealing_title, 'Foo')
-        assert_equal(datestring(a.repeal.date), '2012-02-01')
-
-        assert_validates(a)
-
-        # check that clearing it works
-        a.repeal = None
-        assert_is_none(a.repeal)
-
-        assert_validates(a)
-
-    def test_main(self):
-        a = Act()
-        self.assertEqual(a.main, a.act)
-
-    def test_bad_xml(self):
-        with self.assertRaises(LxmlSyntaxError):
-            Act('badness')
-
-        with self.assertRaises(ValueError):
-            Act('<root>no namespace</root>')
-
-        with self.assertRaises(ValueError):
-            Act('<akomaNtoso>no first child</akomaNtoso>')
-
-        with self.assertRaises(ValueError):
-            Act('<akomaNtoso><child>no namespace</child></akomaNtoso>')
-
-
-class JudgmentTestCase(TestCase):
-    maxDiff = None
-
-    def test_empty_judgment(self):
-        j = Judgment()
-        assert_equal(j.title, "Untitled")
-        assert_is_not_none(j.meta)
-        assert_is_not_none(j.judgmentBody)
-
-    def test_empty_body(self):
-        j = Judgment()
-        assert_not_equal(j.judgmentBody.text, '')
-
-    def test_main(self):
-        j = Judgment()
-        self.assertEqual(j.main, j.judgment)
